requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
WTForms==3.1.2
python-dotenv==1.0.1
requests==2.32.3
email-validator==2.2.0   # en caso de usar EmailField
Flask-Login==0.6.3
SQLAlchemy==2.0.36
PyMySQL==1.1.1
Flask-SQLAlchemy==3.0.3
bcrypt==4.0.1
SQLAlchemy-Utils==0.39.0
Flask-Migrate==4.0.4
alembic==1.11.1
gunicorn==21.2.0
